Removed GET route as this should only be a headless API. Renamed
